About Personal project using Java, MySQL, and JavaFX. An inventory management system for a storefront. Exploring authentication, a custom-built RBAC, and XML & JSON.
This is an inventory management system that uses a JavaFX frontend, a Java API, and MySQL for a database. It allows a store's inventory staff to log in securely with a custom role based access control system. Product inventory, customer invoices, and product category are the main features of this project

Authentication:
This project's authentication system involves a java class called Authenticator that verifies the user's credentials with an XML document.

Authorization:
Authorization in this project is handled by a java class called Authorizer, which is an implementation of a custom role-based access control system. Every CRUD action in this application initiates the Authorizer class to verify that the user has proper privileges. The privilege and role information is stored in a JSON file, which is accessed only by the Authorizer class.

Architecture:
This application uses a Model-View-Controller (MVC) architectural pattern. The models are Java classes that hold the basic data-related logic such as Product, Manufacturer, User, and Customer logic. The views are FXML files holding UI components of each application page. The controllers are Java classes that work as an interface between the View and model components. In addition, there are several helper/service classes that aid the controllers in business-logic